question 4.3

+++ src/Controller/GameController.php
@@ -6,6 +6,7 @@ use App\Entity\Game;
 use App\Entity\UserCharacters;
 use App\Form\GameType;
 use App\Repository\GameRepository;
+use App\Security\AppAccess;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -54,6 +55,8 @@ class GameController extends AbstractController
      */
     public function show(Game $game): Response
     {
+
+        $this->denyAccessUnlessGranted(AppAccess::USERGAME_SHOW, $game);
         return $this->render('game/show.html.twig', [
             'game' => $game,
         ]);
@@ -64,6 +67,7 @@ class GameController extends AbstractController
      */
     public function edit(Request $request, Game $game): Response
     {
+        $this->denyAccessUnlessGranted(AppAccess::USERGAME_EDIT, $game);
         $form = $this->createForm(GameType::class, $game);
         $form->handleRequest($request);
 
@@ -86,6 +90,7 @@ class GameController extends AbstractController
      */
     public function delete(Request $request, Game $game): Response
     {
+        $this->denyAccessUnlessGranted(AppAccess::USERGAME_DELETE, $game);
         if ($this->isCsrfTokenValid('delete'.$game->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($game);
